test: keep CLI execute support discoverable
